+++ src/model/mod.rs
@@ -1,1 +1,2 @@
+#[doc(hidden)]
 pub mod ir;